Moved Network Tests to a different packages
This tests are more like integration tests
